README: unarchive pridefetch, and add sfetch
The real owner (@ megabytesofrem) of pridefetch seems to deleted
their repo and now it resides on a fork.

While searching for sfetch (in ARCHIVE.md), I encounter with
similar "fetch" scripts and saw one of them (with exactly same
name, called sfetch) wasn't here, so I added it as well.